+++ symfony/src/Controller/ArtistController.php
@@ -7,35 +7,41 @@ use Symfony\Component\{
     HttpFoundation\Response,
     Routing\Annotation\Route
 };
-use App\Entity\AddScene;
-use App\Entity\User;
+use App\Entity\{
+    AddScene,
+    User
+};
 use App\Form\AddSceneType;
 use App\Repository\AddSceneRepository;
+use App\Service\{
+    CodeFileUploader,
+    OptionFieldProcessor
+};
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\Filesystem\Filesystem;
 
 #[Route("/artist", name: "artist_")]
 class ArtistController extends AbstractController
 {
 
     #[Route("/add-new-scene/{id}", name: "addScene", methods: ["GET", "POST"])]
-    public function addScene( EntityManagerInterface $entityManager, Request $request, ?int $id ): Response
-    {   
+    public function addScene( 
+        EntityManagerInterface $entityManager, 
+        Request $request, 
+        CodeFileUploader $uploader,
+        OptionFieldProcessor $optionField,
+        ?int $id 
+    ): Response {   
         $sceneRequest = new AddScene(); 
 
         $form = $this->createForm(AddSceneType::class, $sceneRequest);
         $form -> handleRequest($request); 
      
         if ( $form->isSubmitted() && $form->isValid()){
-            /** @var UploadedFile $file */
+
             $file = $form->get('codeFile')->getData();
-            // Extract the file name and extension
-            $fileName = pathinfo($file->getClientOriginalName(), PATHINFO_FILENAME);
-            $fileExtension = $file->guessExtension();           
-            // Generate a unique file name
-            $uniqueFileName = $fileName. uniqid(). '.'. $fileExtension;
-                        
-            $file->move($this->getParameter('code_directory'), $uniqueFileName);
+            $uniqueFileName = $uploader->upload($file);
             $sceneRequest->setCodeFile($uniqueFileName);
         
             $sceneRequest->setUser($this->getUser());
@@ -43,11 +49,9 @@ class ArtistController extends AbstractController
             // Add the value of the "Other" option to the main language array
             $language = $sceneRequest->getLanguage();
             $otherLanguage = $form->get('otherLanguage')->getData();
-           
-            if ($otherLanguage && in_array("other", $language)) {
-                $language[] = $otherLanguage;
-            }
-            $sceneRequest->setLanguage($language);
+            $mergedLanguages = $optionField->appendCustomChoice($language, $otherLanguage);
+            $sceneRequest->setLanguage($mergedLanguages);
+
             $entityManager->persist($sceneRequest);
             $entityManager->flush();
 
@@ -71,7 +75,6 @@ class ArtistController extends AbstractController
             throw $this->createAccessDeniedException('You must be logged in to access this page.');
         }
         $userId = $user->getId();
-        $sceneRequests = $newScene->findby(['user' => $userId]);
         // Sorting all requests by creation date
         $sceneRequests = $newScene->findBy(
             ['user' => $userId],
@@ -82,12 +85,23 @@ class ArtistController extends AbstractController
            'sceneRequests' => $sceneRequests,
         ]);
     }
-    
 
     #[Route("/delete/request/{id}", name: "delete_request", methods: ["GET", "POST"])]
     public function deleteSceneRequest(EntityManagerInterface $entityManager, AddSceneRepository $sceneReq, $id): Response
     {
         $request = $sceneReq->find($id);
+        if (!$request) {
+            throw $this->createNotFoundException('Scene not found.');
+        }
+        $user = $this->getUser();
+        if (!$user || $request->getUser() !== $user) {
+            throw $this->createAccessDeniedException('You are not allowed to delete this request.');
+        }
+        $filesystem = new Filesystem();
+        $filePath = $this->getParameter('code_directory') . '/' . $request->getCodeFile();
+        if ($filesystem->exists($filePath)) {
+            $filesystem->remove($filePath);
+        }
         $entityManager->remove($request);
         $entityManager->flush();
         $this->addFlash('success', 'Request successfully deleted');
